requirements.txt
# Core dependencies
opencv-contrib-python>=4.8.0
numpy>=1.21.0
typer>=0.9.0
PyYAML>=6.0
